refactor + add shot by click on sector
