+++ step_08/App.tsx
@@ -7,7 +7,7 @@ import { IPoint } from "./DataModel";
 export class App extends Component<{},{}> {
     //shots = [{from:{x:0,y:0}, to:{x:100, y:200}}, {from:{x:100,y:200}, to:{x:200, y:50}}];
     shotListModel: ShotListModel = new ShotListModel;
-    static offSet: IPoint = {x: 663641, y: 7455053}
+    static offSet: IPoint = {x: 663641, y: 7455054}
     render(){
         return (
             <div id="container">
