+++ step_08/MapComponent.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
 import { Component } from 'react';
-import { observable, action } from 'mobx';
+import { observable } from 'mobx';
 import { observer } from "mobx-react";
-import { IShot, IData, IPoint } from "./DataModel";
 import { Areas } from "./Areas";
 import { LineGrid } from "./LineGrid";
 import { PointGrid } from "./PointGrid"
-import { ShotListModel, ShotModel } from "./ShotList";
-import { Data } from "./FetchData";
+import { ShotListModel } from "./ShotList";
+import { Data } from "./LoadData";
 import { App } from "./App";
+import { ShotComponent } from "./ShotComponent";
 
 export interface MapComponentProps{
     shotList: ShotListModel;
@@ -50,39 +50,4 @@ export class MapComponent extends Component<MapComponentProps, {}>{
             </div>
         );
     } 
-}
-
-export interface ShotProps{
-    shot: ShotModel;
-    padding: number;
-    shots: ShotListModel;
-}
-
-@observer
-export class ShotComponent extends Component<ShotProps, {}>{
-    render(){
-        const shot = this.props.shot;
-        const padding = this.props.padding;
-        const textX = padding + (shot.from.point.x + shot.to.point.x)/2 - 2;
-        const textY = padding + (shot.from.point.y + shot.to.point.y)/2 - 2;
-        const textDegrees = Math.atan((shot.from.point.y - shot.to.point.y)/(shot.from.point.x - shot.to.point.x)) * 180 / Math.PI;
-        
-        return(
-            <g id="group">
-                <path stroke="red" strokeWidth="2" 
-                    d={"M " + (shot.from.point.x + padding) + " " + (shot.from.point.y + padding) + "L" + (shot.to.point.x + padding) + " " + (shot.to.point.y + padding)} 
-                />
-                <circle className="draggable" stroke="black" 
-                    onMouseDown={() => this.props.shots.startMoving(shot.from)}
-                    cx={shot.from.point.x + padding} cy={shot.from.point.y + padding} r="3" />
-                <circle className="draggable" stroke="black"
-                    onMouseDown={() => this.props.shots.startMoving(shot.to)} 
-                    cx={shot.to.point.x + padding} cy={shot.to.point.y + padding} r="3" />
-                <text transform={"rotate(" + textDegrees + " " + textX + "," + textY + ")"} 
-                    x={textX}
-                    y={textY} fontSize="10">{shot.key}
-                </text>
-            </g>
-        )
-    }
 }
